+++ src/spring/azext_spring/custom.py
@@ -91,7 +91,8 @@ def _update_application_insights_asc_create(cmd,
 def spring_update(cmd, client, resource_group, name, app_insights_key=None, app_insights=None,
                   disable_app_insights=None, sku=None, tags=None, build_pool_size=None,
                   enable_log_stream_public_endpoint=None, enable_dataplane_public_endpoint=None,
-                  ingress_read_timeout=None, enable_planned_maintenance=False, planned_maintenance_day=None,
+                  enable_private_storage_access=None, ingress_read_timeout=None, 
+                  enable_planned_maintenance=False, planned_maintenance_day=None,
                   planned_maintenance_start_hour=None, no_wait=False):
     """
     TODO (jiec) app_insights_key, app_insights and disable_app_insights are marked as deprecated.
@@ -102,6 +103,7 @@ def spring_update(cmd, client, resource_group, name, app_insights_key=None, app_
     update_service_tags = False
     update_service_sku = False
     update_dataplane_public_endpoint = False
+    update_private_storage_access = False
 
     # update service sku
     if sku is not None:
@@ -114,15 +116,24 @@ def spring_update(cmd, client, resource_group, name, app_insights_key=None, app_
     updated_resource_properties.zone_redundant = None
 
     if enable_log_stream_public_endpoint is not None or enable_dataplane_public_endpoint is not None:
+        if updated_resource_properties.vnet_addons is None:
+            updated_resource_properties.vnet_addons = models.ServiceVNetAddons()
         val = enable_log_stream_public_endpoint if enable_log_stream_public_endpoint is not None else \
             enable_dataplane_public_endpoint
-        updated_resource_properties.vnet_addons = models.ServiceVNetAddons(
-            data_plane_public_endpoint=val,
-            log_stream_public_endpoint=val
-        )
+        updated_resource_properties.vnet_addons.data_plane_public_endpoint = val
+        updated_resource_properties.vnet_addons.log_stream_public_endpoint = val
         update_dataplane_public_endpoint = True
-    else:
-        updated_resource_properties.vnet_addons = None
+
+    if enable_private_storage_access is not None:
+        if updated_resource_properties.vnet_addons is None:
+            updated_resource_properties.vnet_addons = models.ServiceVNetAddons(
+                # explicitly set as none in case unexpected update
+                data_plane_public_endpoint = None,
+                log_stream_public_endpoint = None
+            )
+        val = "Enabled" if enable_private_storage_access else "Disabled"
+        updated_resource_properties.vnet_addons.private_storage_access = val
+        update_private_storage_access = True
 
     _update_application_insights_asc_update(cmd, resource_group, name, location,
                                             app_insights_key, app_insights, disable_app_insights, no_wait)
@@ -146,7 +157,7 @@ def spring_update(cmd, client, resource_group, name, app_insights_key=None, app_
         update_service_tags = True
 
     if update_service_tags is False and update_service_sku is False and update_dataplane_public_endpoint is False \
-            and update_planned_maintenance is False and ingress_read_timeout is None:
+            and update_private_storage_access is False and update_planned_maintenance is False and ingress_read_timeout is None:
         return resource
 
     updated_resource.properties = updated_resource_properties